[FIX] MAKE D=4 ok sauf pour core_free

+++ corewar_dir/srcs/ope/get_args.c
@@ -19,24 +19,24 @@ static int32_t			find_arg(t_carriages *current, t_data *data,
 
 	if (mask == T_REG)
 	{
-		arg = core_get_reg(data, current->position + current->to_jump, current);
+		arg = core_get_reg(data, (int16_t)(current->position + current->to_jump), current);
 		if (arg == FAILURE)
 			*flag |= BAD_REG_NB;
 		else if ((*flag & REG_NUM) == FALSE)
 		{
 			*flag = GET;
-			arg = set_reg_value(current, arg, NO_NEED, (uint8_t *)flag);
+			arg = set_reg_value(current, (int8_t)arg, NO_NEED, (uint8_t *)flag);
 		}
 		current->to_jump += MEM_REG;
 	}
 	else if (mask == T_DIR)
 	{
-		arg = core_get_dir(data, current->position + current->to_jump, *flag);
+		arg = core_get_dir(data, (int16_t)(current->position + current->to_jump), *flag);
 		current->to_jump += (*flag & SMALL_DIR) ? MEM_SMALL_DIR : MEM_DIR;
 	}
 	else
 	{
-		arg = core_get_ind(data, current->position, current->to_jump, *flag);
+		arg = core_get_ind(data, (int16_t)current->position, (int16_t)current->to_jump, *flag);
 		current->to_jump += MEM_IND;
 	}
 	return (arg);
@@ -76,7 +76,7 @@ static uint8_t			get_cur_ocp(t_carriages *current, t_data *data,
 {
 	uint8_t	ocp;
 
-	ocp = core_get_ocp(data, current->position + current->to_jump);
+	ocp = core_get_ocp(data, (int16_t)(current->position + current->to_jump));
 	current->to_jump += MEM_OCP;
 	if (ocp_verification(ocp, *type) == FAILURE)
 		*type = NO_OP;
